ml_service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0

tensorflow==2.15.0
keras==2.15.0

numpy==1.26.4
pandas==2.0.3
Pillow==10.2.0
scikit-learn==1.3.0
scipy==1.11.4

kaggle==1.6.14
requests==2.31.0
python-dotenv==1.0.0
